requirements.txt
plotapi==1.1.0
arrow==0.17.0
streamlit==0.85.1
~umpy==1.18.5
pandas==1.0.5
requests==2.24.0
bar_chart_race==0.1.0
plotly==4.14.3
fintoc==0.3.1
numpy==1.21.1
Pillow==8.3.1
